+++ package/lib/src/controls/tabs.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:flet/src/utils/alignment.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
@@ -75,6 +76,9 @@ class _TabsControlState extends State<TabsControl>
   Widget build(BuildContext context) {
     debugPrint("TabsControl build: ${widget.control.id}");
 
+    // keep only visible tabs
+    widget.children.retainWhere((c) => c.isVisible);
+
     var tabs = StoreConnector<AppState, ControlsViewModel>(
         distinct: true,
         converter: (store) => ControlsViewModel.fromStore(
@@ -121,12 +125,12 @@ class _TabsControlState extends State<TabsControl>
 
           var indicatorBorderRadius =
               parseBorderRadius(widget.control, "indicatorBorderRadius");
-          var inidicatorBorderSide = parseBorderSide(
-              Theme.of(context), widget.control, "inidicatorBorderSide");
+          var indicatorBorderSide = parseBorderSide(
+              Theme.of(context), widget.control, "indicatorBorderSide");
           var indicatorPadding =
               parseEdgeInsets(widget.control, "indicatorPadding");
 
-          var inidicatorColor = HexColor.fromString(Theme.of(context),
+          var indicatorColor = HexColor.fromString(Theme.of(context),
                   widget.control.attrString("indicatorColor", "")!) ??
               TabBarTheme.of(context).indicatorColor ??
               Theme.of(context).colorScheme.primary;
@@ -136,9 +140,14 @@ class _TabsControlState extends State<TabsControl>
 
           var indicatorTabSize = widget.control.attrBool("indicatorTabSize");
 
+          var isScrollable = widget.control.attrBool("scrollable", true)!;
+          var tabAlignment = parseTabAlignment(widget.control, "tabAlignment",
+              isScrollable ? TabAlignment.start : TabAlignment.fill);
+
           var tabBar = TabBar(
+              tabAlignment: tabAlignment,
               controller: _tabController,
-              isScrollable: widget.control.attrBool("scrollable", true)!,
+              isScrollable: isScrollable,
               dividerColor:
                   HexColor.fromString(Theme.of(context), widget.control.attrString("dividerColor", "")!) ??
                       TabBarTheme.of(context).dividerColor,
@@ -148,7 +157,7 @@ class _TabsControlState extends State<TabsControl>
                       : TabBarIndicatorSize.label)
                   : TabBarTheme.of(context).indicatorSize,
               indicator: indicatorBorderRadius != null ||
-                      inidicatorBorderSide != null ||
+                      indicatorBorderSide != null ||
                       indicatorPadding != null
                   ? UnderlineTabIndicator(
                       borderRadius: indicatorBorderRadius ??
@@ -156,17 +165,17 @@ class _TabsControlState extends State<TabsControl>
                           const BorderRadius.only(
                               topLeft: Radius.circular(2),
                               topRight: Radius.circular(2)),
-                      borderSide: inidicatorBorderSide ??
+                      borderSide: indicatorBorderSide ??
                           themeIndicator?.borderSide ??
                           BorderSide(
                               width: themeIndicator?.borderSide.width ?? 2,
                               color: themeIndicator?.borderSide.color ??
-                                  inidicatorColor),
+                                  indicatorColor),
                       insets: indicatorPadding ??
                           themeIndicator?.insets ??
                           EdgeInsets.zero)
                   : TabBarTheme.of(context).indicator,
-              indicatorColor: inidicatorColor,
+              indicatorColor: indicatorColor,
               labelColor: HexColor.fromString(Theme.of(context), widget.control.attrString("labelColor", "")!) ??
                   TabBarTheme.of(context).labelColor ??
                   Theme.of(context).colorScheme.primary,