+++ sdk/python/packages/flet-core/src/flet_core/tabs.py
@@ -14,6 +14,7 @@ from flet_core.types import (
     ResponsiveNumber,
     RotateValue,
     ScaleValue,
+    TabAlignment,
 )
 
 
@@ -25,8 +26,9 @@ class Tab(Control):
         tab_content: Optional[Control] = None,
         ref: Optional[Ref] = None,
         icon: Optional[str] = None,
+        visible: Optional[bool] = None,
     ):
-        Control.__init__(self, ref=ref)
+        Control.__init__(self, ref=ref, visible=visible)
         self.text = text
         self.icon = icon
         self.__content: Optional[Control] = None
@@ -162,6 +164,7 @@ class Tabs(ConstrainedControl):
         tabs: Optional[List[Tab]] = None,
         selected_index: Optional[int] = None,
         scrollable: Optional[bool] = None,
+        tab_alignment: Optional[TabAlignment] = None,
         animation_duration: Optional[int] = None,
         divider_color: Optional[str] = None,
         indicator_color: Optional[str] = None,
@@ -206,6 +209,7 @@ class Tabs(ConstrainedControl):
         self.tabs = tabs
         self.selected_index = selected_index
         self.scrollable = scrollable
+        self.tab_alignment = tab_alignment
         self.animation_duration = animation_duration
         self.divider_color = divider_color
         self.label_color = label_color
@@ -266,6 +270,16 @@ class Tabs(ConstrainedControl):
     @scrollable.setter
     def scrollable(self, value: Optional[bool]):
         self._set_attr("scrollable", value)
+
+    # tab_alignment
+    @property
+    def tab_alignment(self) -> Optional[TabAlignment]:
+        return self.__tab_alignment
+
+    @tab_alignment.setter
+    def tab_alignment(self, value: Optional[TabAlignment]):
+        self.__tab_alignment = value
+        self._set_attr("tabAlignment", value.value if value is not None else None)
 
     # animation_duration
     @property