+++ sdk/python/packages/flet-core/tests/test_navigation_bar.py
@@ -20,13 +20,13 @@ def test_instance_no_attrs_set():
 def test_extension_set_enum():
     r = ft.NavigationBar()
     assert r.label_behavior is None
-    assert r._get_attr("labelType") is None
+    assert r._get_attr("labelBehavior") is None
 
     r = ft.NavigationBar(label_behavior=ft.NavigationBarLabelBehavior.ALWAYS_SHOW)
     assert isinstance(r.label_behavior, ft.NavigationBarLabelBehavior)
     assert r.label_behavior == ft.NavigationBarLabelBehavior.ALWAYS_SHOW
-    assert r._get_attr("labelType") == "alwaysShow"
+    assert r._get_attr("labelBehavior") == "alwaysShow"
 
     r = ft.NavigationBar(label_behavior="alwaysHide")
     assert isinstance(r.label_behavior, str)
-    assert r._get_attr("labelType") == "alwaysHide"
+    assert r._get_attr("labelBehavior") == "alwaysHide"